+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change examples to use `FooStorage<Self>` and destructure system data in method head ([#206])
 * `AntiStorage` for `ChecStorage` ([#208])
 * Integrate futures by introducing a `common` module ([#209])
+* Add lazy component insertions ([#214])
 
 [#198]: https://github.com/slide-rs/specs/pull/198
 [#199]: https://github.com/slide-rs/specs/pull/199
@@ -12,3 +13,4 @@
 [#206]: https://github.com/slide-rs/specs/pull/206
 [#208]: https://github.com/slide-rs/specs/pull/208
 [#209]: https://github.com/slide-rs/specs/pull/209
+[#214]: https://github.com/slide-rs/specs/pull/214